separate contagionRiskInput into new component
