+++ src/components/Rules/AddRuleForm/components/ContagionRiskInput/styles.js
@@ -1,0 +1,19 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  formControlContagionRisk: {
+    margin: theme.spacing(4),
+    width: '400px',
+  },
+  labelContagionRisk: {
+    marginLeft: '10px',
+  },
+  gridContagionRisk: {
+    textAlign: 'center',
+  },
+  optionsContagionRisk: {
+    justifyContent: 'center',
+  },
+}));
+
+export default useStyles;
